Move assert_that, equal_to, is_empty to apache_beam.testing.util
