+++ sdks/python/apache_beam/testing/util_test.py
@@ -15,13 +15,13 @@
 # limitations under the License.
 #
 
-"""Unit tests for the util transforms."""
+"""Unit tests for testing utilities."""
 
 import unittest
 
 from apache_beam import Create
 from apache_beam.testing.test_pipeline import TestPipeline
-from apache_beam.transforms.util import assert_that, equal_to, is_empty
+from apache_beam.testing.util import assert_that, equal_to, is_empty
 
 
 class UtilTest(unittest.TestCase):
